address review feedback on dispatch counter coverage
