+++ projects/rocprofiler-sdk/tests/rocprofv3/advanced-thread-trace/CMakeLists.txt
@@ -206,13 +206,14 @@ rocprofiler_add_integration_execute_test(
 
 rocprofiler_add_integration_validate_test(
     rocprofv3-test-att-hsa-multiqueue-plus-pmc
-    TEST_PATHS validate_counter_collection.py
+    TEST_PATHS validate.py
     COPY conftest.py
     CONFIG pytest.ini
     TIMEOUT 45
     LABELS "integration-tests"
     DISABLED ${IS_DISABLED}
     FIXTURES_REQUIRED rocprofv3-test-att-hsa-multiqueue-plus-pmc
+    DISCOVERY_ARGS -k test_counter_collection_with_att
     ARGS --input ${CMAKE_CURRENT_BINARY_DIR}/vector-ops-trace/cmd_input/out_results.json
     FAIL_REGULAR_EXPRESSION "AssertionError")
