--- projects/rocprofiler-sdk/tests/rocprofv3/counter-collection/invalid-options/conftest.py
@@ -1,44 +1,0 @@
-#!/usr/bin/env python3
-
-# MIT License
-#
-# Copyright (c) 2026 Advanced Micro Devices, Inc. All rights reserved.
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.  IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-# THE SOFTWARE.
-
-import pytest
-
-
-def pytest_addoption(parser):
-    parser.addoption("--rocprofv3", action="store", help="Path to rocprofv3.")
-    parser.addoption(
-        "--pmc-groups-input",
-        action="store",
-        help="Path to the input file containing pmc_groups.",
-    )
-
-
-@pytest.fixture(scope="session")
-def rocprofv3(request):
-    return request.config.getoption("--rocprofv3")
-
-
-@pytest.fixture(scope="session")
-def pmc_groups_input(request):
-    return request.config.getoption("--pmc-groups-input")
