--- projects/rocprofiler-sdk/tests/rocprofv3/counter-collection/invalid-options/pytest.ini
@@ -1,5 +1,0 @@
-
-[pytest]
-addopts = --durations=20 -rA -s
-testpaths = test_invalid_options.py
-pythonpath = @ROCPROFILER_SDK_TESTS_BINARY_DIR@/pytest-packages
